requirements.txt
pytest-bdd>=8.1.0
setuptools>=80.9.0
selenium>=4.35.0
pytest>=8.4.1
python-dotenv>=1.1.1
allure-python-commons>=2.15.0
allure-pytest>=2.15.0
pytest-failed-screenshot==1.0.2